aplicacao/backend/requirements.txt
psycopg[binary]
python-dotenv
seaborn
google-generativeai
tabulate
matplotlib